requirements.txt
streamlit==1.52.2
openai==2.14.0
st-chat-message==0.3.8
dotenv==0.9.9
pycountry==24.6.1
pytz==2025.2
 pyyaml==6.0.3